python/requirements.txt
matplotlib==3.6.2
numpy==1.19.2
Pillow==10.4.0
torch==2.0.0+cu118
torchvision==0.15.1+cu118
tqdm==4.66.2
python ==3.8